move to own file, rename columns
